+++ src/controllers/course-controller.js
@@ -2,6 +2,8 @@ const { StatusCodes } = require('http-status-codes');
 const { SuccessResponse , ErrorResponse } = require('../utils/common');
 const AppError = require('../utils/error/app-error');
 const {CourseService} = require('../services');
+const Bootcamp = require('../models/bootcamps');
+const Course = require('../models/courses');
 
 async function getCourses(request,response){
     try {
@@ -48,6 +50,19 @@ async function postCourses(request,response){
     try {
         request.body.bootcamp = request.params.bootcampId;
         console.log('request body is',request.body);
+        request.body.user = request.user.id;
+        //checking if the bootcamp exists for which course is getting created
+        const bootcamp = await Bootcamp.findById(request.params.bootcampId);
+        if(!bootcamp){
+            throw new AppError(`No Bootcamp exist with the given Id ${request.params.bootcampId}`,StatusCodes.NOT_FOUND);
+        }
+
+
+        //check if the bootcamp owner is creating the course or not
+        if(bootcamp.user.toString() !== request.user.id && request.user.role!=='admin'){
+            throw new AppError(`User ${request.user.id} is not authorized to add this Course to this Bootcamp`,StatusCodes.UNAUTHORIZED);
+        }
+
         const course = await CourseService.postCourses(request.body);
         if(!course){
             throw new AppError(`Something went wrong while creating course`,StatusCodes.BAD_REQUEST);
@@ -64,10 +79,17 @@ async function postCourses(request,response){
 
 async function updateCourseById(request,response){
     try {
-        const course = await CourseService.putCoursesById(request);
+        let course = await Course.findById(request.params.id);
         if(!course){
             throw new AppError(`No Course Exists for the given CourseId`,StatusCodes.OK);
         }
+        //make sure that user is course owner 
+        if(course.user.toString() !== request.user.id && request.user.role!=='admin'){
+            throw new AppError(`User ${request.user.id} is not authorized to update this Course to this Bootcamp`,StatusCodes.UNAUTHORIZED);
+        }
+
+        course = await CourseService.putCoursesById(request);
+        
         SuccessResponse.data = course;
         return response.status(StatusCodes.OK).json(SuccessResponse);
     } catch (error) {
@@ -81,10 +103,16 @@ async function updateCourseById(request,response){
 async function deleteCourseById(request,response){
     try {
         console.log(' in deleteCourseById')
-        const course = await CourseService.deleteCoursesById(request.params.id);
+        let course = await Course.findById(request.params.id);
         if(!course){
             throw new AppError(`No Course Exists for the given CourseId`,StatusCodes.OK);
         }
+        //make sure that user is course owner 
+        if(course.user.toString() !== request.user.id && request.user.role!=='admin'){
+            throw new AppError(`User ${request.user.id} is not authorized to update this Course to this Bootcamp`,StatusCodes.UNAUTHORIZED);
+        }
+
+        course =  await CourseService.deleteCoursesById(request.params.id);
         SuccessResponse.data = course;
         return response.status(StatusCodes.OK).json(SuccessResponse);
     } catch (error) {
